Probably production-ready. Prepairing for Task Scheduler
Also now using Hyper-V core. As a side note: while the new Server Manager console is pretty slick, remote management of Windows has either been largely ignored by Microsoft for the majority of their time in the industry, or I'm doing it wrong.
It's probably both.
I really enjoy SSH and an OS where everything can be done in the command line if need be.