Many things! Fail if low space & email if too fast
A reasonably big update, so I figured it deserved a major increment. The free space check is terrible, but probably functional. I haven't actually tested it. Recently a few disk2vhd attempts have silently failed due to low disk space, so I added the space check and the elapsed time check both fire off an email if they fail.

Thankfully, the cleanup script triggers on the number of files in a directory, and not the amount of space available, so no harm came of missing a few disk2vhd triggers.