Mediocre Hyper-V backup script
Line 14 really leverages the benefits of both PowerShell and Hyper-V. Line 12 is on-par with Bash and GNU coreutils. Every other line of code has been included to work around the many shortcomings of PowerShell, Hyper-V, or Windows. It took me 3 days to write 8 lines, and a 4th day to Windows Task Scheduler working. PowerShell is great, and just needs a few more years of maturation. Windows will always be a stick in the mud. I hope Windows Server keeps losing market share so Microsoft will continue to improve PowerShell on Linux.